feat(billing): differentiate top-up webhook + surface credits on assistants page
- Add IsTopUpProductID() to billing.Repository so the webhook handler can
  recognise one-time credit-pack orders without coupling to polar.Catalog.
- HandlePolarWebhook forks order.paid + top-up product into
  handleTopUpOrder: invalidate caches, refresh period usage, emit a new
  gram_topup_purchased posthog event. Skips the subscription tier-refresh
  + OpenRouter limit refresh + gram_subscription_changed event, which are
  all no-ops for top-ups and would otherwise pollute analytics.
- Drop the unused costToCredits helper + metadata["credits"] write. The
  Polar Chat Credits meter aggregates sum(cost), not the integer credit
  field, so emitting credits did nothing. Aligns the code with the
  current meter configuration.
- Extract TopUpCTA + UsageProgress out of Billing.tsx into
  components/billing/usage-controls.tsx for reuse.
- Assistants index renders an Assistant Credits section above the list:
  shows credits used / included with the same progress UI as the billing
  page, plus an admin-gated TOP UP CREDITS button. Assistants are
  headless so users wouldn't see an inline 402; the assistants page is
  the natural surface to show remaining balance and the top-up CTA.
